Move child information component and context
