+++ frontend/src/employee-frontend/components/child-information/ChildInformation.tsx
@@ -25,27 +25,27 @@ import { fontWeights } from 'lib-components/typography'
 import { Gap } from 'lib-components/white-space'
 import { faUsers } from 'lib-icons'
 
-import { ChildContext, ChildContextProvider, ChildState } from '../state/child'
-import { useTranslation } from '../state/i18n'
-import { TitleContext, TitleState } from '../state/title'
-import { UserContext } from '../state/user'
+import { useTranslation } from '../../state/i18n'
+import { TitleContext, TitleState } from '../../state/title'
+import { UserContext } from '../../state/user'
+import CircularLabel from '../common/CircularLabel'
+import WarningLabel from '../common/WarningLabel'
+import { getLayout, Layouts } from '../layouts'
 
-import Assistance from './child-information/Assistance'
-import BackupCare from './child-information/BackupCare'
-import ChildApplications from './child-information/ChildApplications'
-import ChildDetails from './child-information/ChildDetails'
-import ChildDocumentsSection from './child-information/ChildDocumentsSection'
-import ChildIncome from './child-information/ChildIncome'
-import DailyServiceTimesSection from './child-information/DailyServiceTimesSection'
-import FamilyContacts from './child-information/FamilyContacts'
-import FeeAlteration from './child-information/FeeAlteration'
-import GuardiansAndParents from './child-information/GuardiansAndParents'
-import PedagogicalDocuments from './child-information/PedagogicalDocuments'
-import Placements from './child-information/Placements'
-import ServiceApplicationsSection from './child-information/ServiceApplicationsSection'
-import CircularLabel from './common/CircularLabel'
-import WarningLabel from './common/WarningLabel'
-import { getLayout, Layouts } from './layouts'
+import Assistance from './Assistance'
+import BackupCare from './BackupCare'
+import ChildApplications from './ChildApplications'
+import ChildDetails from './ChildDetails'
+import ChildDocumentsSection from './ChildDocumentsSection'
+import ChildIncome from './ChildIncome'
+import DailyServiceTimesSection from './DailyServiceTimesSection'
+import FamilyContacts from './FamilyContacts'
+import FeeAlteration from './FeeAlteration'
+import GuardiansAndParents from './GuardiansAndParents'
+import PedagogicalDocuments from './PedagogicalDocuments'
+import Placements from './Placements'
+import ServiceApplicationsSection from './ServiceApplicationsSection'
+import { ChildContext, ChildContextProvider, ChildState } from './state'
 
 const HeaderRow = styled.div`
   display: flex;
